+++ packages/real-time/CHANGELOG.md
@@ -1,3 +1,3 @@
-# @the-forgebase/real-time
+# @forgebase/real-time
 
 ## 0.0.0